+++ dolibarr/custom/novoux/theme/novo/style.css.php
@@ -98,7 +98,36 @@ if (defined('THEME_ONLY_CONSTANT')) {
 
 session_cache_limiter('public');
 
-require_once __DIR__.'/../../main.inc.php'; // __DIR__ allow this script to be included in custom themes
+// Load Dolibarr environment. This file lives at <module>/theme/novo/, so it must
+// resolve main.inc.php for both supported install roots: htdocs/custom/novoux/
+// and htdocs/novoux/ (Dolibarr packaging rule). Mirrors css/novo-inject.css.php.
+$res = 0;
+if (!$res && !empty($_SERVER["CONTEXT_DOCUMENT_ROOT"])) {
+	$res = @include $_SERVER["CONTEXT_DOCUMENT_ROOT"]."/main.inc.php";
+}
+$tmp = empty($_SERVER['SCRIPT_FILENAME']) ? '' : $_SERVER['SCRIPT_FILENAME'];
+$tmp2 = realpath(__FILE__);
+$i = strlen($tmp) - 1;
+$j = strlen($tmp2) - 1;
+while ($i > 0 && $j > 0 && isset($tmp[$i]) && isset($tmp2[$j]) && $tmp[$i] == $tmp2[$j]) {
+	$i--;
+	$j--;
+}
+if (!$res && $i > 0 && file_exists(substr($tmp, 0, ($i + 1))."/main.inc.php")) {
+	$res = @include substr($tmp, 0, ($i + 1))."/main.inc.php";
+}
+if (!$res && $i > 0 && file_exists(dirname(substr($tmp, 0, ($i + 1)))."/main.inc.php")) {
+	$res = @include dirname(substr($tmp, 0, ($i + 1)))."/main.inc.php";
+}
+if (!$res && file_exists("../../../main.inc.php")) {
+	$res = @include "../../../main.inc.php";
+}
+if (!$res && file_exists("../../../../main.inc.php")) {
+	$res = @include "../../../../main.inc.php";
+}
+if (!$res) {
+	die("Include of main fails");
+}
 // From main.inc.php
 '
 @phan-var-force MenuManager $menumanager